+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * New Features:
     * `faspex5`: Support public and private invitations.
+    * `shares`: Added group membership management.
   * Issues Fixed:
     * `aoc`: #135 `exclude_dropbox_packages` query option can be overridden
   * Breaking Changes: